added customer vehicle search, working on testing

+++ src/ca/ubc/cs304/model/VehicleModel.java
@@ -14,13 +14,13 @@ public class VehicleModel {
     private final String model;
     private final int year;
     private final String color;
-    private final Float odometer;
+    private final Double odometer;
     private final Status status;
     private final String vtname; // TODO: figure out if I should exchange for VehicleTypeObject instead
     private final String location; // TODO: figure out if for location and city should be Branch object but keep simple for now
     private final String city;
 
-    public VehicleModel(String vlicense, String make, String model, int year, String color, Float odometer,
+    public VehicleModel(String vlicense, String make, String model, int year, String color, Double odometer,
                         Status status, String vtname, String location, String city) {
         //this.vid = vid;
         this.vlicense = vlicense;
@@ -71,7 +71,7 @@ public class VehicleModel {
         return location;
     }
 
-    public Float getOdometer() {
+    public Double getOdometer() {
         return odometer;
     }
